Make ImapQueue more robust against connection loss
This will reconnect on every polling cycle as this was easier to
implement than only reconnecting on a lost connection. However, for my
use case there is only a low volume of incoming messages, so that it
makes sense to have a long polling interval in which case individual
connections to poll might be even better than a persistent connection.

Due to limitations in the aioimaplib [1], a timeout in one of the
commands will leave a cancelled, but pending task around. This also,
produces warnings in the tests that cannot be silenced.

[1]: https://github.com/bamthomas/aioimaplib/issues/58