Parse IMAP FETCH response properly
in order to support a wider variety of IMAP servers. In particular, the
order of requested pairs is not necessarily deterministic (at least not
across servers). Also, we might get back more than requested (usually
only the FLAGS it seems) and thus need to be able to parse any valid
response field. The body structure isn't fully parsed to keep things at
least somewhat simple (because we don't make use of it). To do so, the
parser make use of the fact that everything will within nested
parenthesized lists and other parenthesis have to be within strings.