Handle non-ascii octets in IMAP messages
This is a bad hack by using latin1 as an encoding where each octet is
decoded to exactly one string character so that the decoded string will
have the same length as the bytes. This change only covers RFC822 fields
and other fields can still give problems if they contain octets > 127.

It is only intended as temporary solution until the parser has been
replaced with something able to handle bytes instead of strings.